Initial 1.5

+++ Source/Source/Core/Harmony Patches/SCPHarmony_Patches.cs
@@ -85,7 +85,7 @@ namespace Foundation.HarmonyPatches
             return false;
         }
 
-        public static void SCP939_VoicesMovementSpeed(bool diagonal, Pawn __instance, ref int __result)
+        public static void SCP939_VoicesMovementSpeed(bool diagonal, Pawn __instance, ref float __result)
         {
             if (__instance.TryGetComp<CompVoices>() == null)
                 return;
